Allow multiple records per model with toSearchableRecords()
Sometimes there are cases where multiple, customized search index records need
to be created per model record. Before this commit, scout-extended only
supported customizing attributes for a single search index record via the
toSearchableArray() method, or morphing one or more of these attributes via the
split*() methods.

This commit adds support for defining a toSearchableRecords() function on the
Eloquent model. This function should return an array of arrays, each array
being a search index record for that model.

    function toSearchableRecords()
    {
        return [
            [
                'title' => $this->title . ' min',
                'temp' => $this->temp_min,
            ],
            [
                'title' => $this->title . ' max',
                'temp' => $this->temp_max,
            ],
        ];
    }

If the toSearchableRecords() method is defined, then the toSearchableArray()
and split*() methods will not be used.